Initial Upload
Initial Upload, not confirmed to be working or anything, just something to get up. Manual mode not yet implemented.

Just using for statements and such until a better solution is found. Currently investigation Newton's Law of Cooling and such. Dr. Simpson ,(math professor in DEs and PDEs) said for what I've described it's gonna more like for statements.

Currently checks when temp is within +/-5 degrees of goal, and adjusts accordingly. Has wait times which will need to be adjusted for water propagation. 

Not tested on actual Raspberry Pi with GPIO's so I do not know if all variables are set accordingly.